server: 🎉 Last pass on the server for LLRP changes. Working!
There are many changes in the last version. Here's a run down... 1) Messages are now relayed through from LLRP servers without having to use the Get_Report message. 2) The readerMap is now properly mapping (this broke after the readerMap changes) 3) A lot of the logging is reduced. In the future, it owuld be nice to filter it by type (debugging, useful info, errors, etc.) instead of just one long list that becomes unweildy. It would also be nice to move the constants somewhere else.